+++ app/src/main/java/com/nabil/rateme/model/Repository.kt
@@ -6,5 +6,5 @@ interface Repository {
 
     fun insertAllMovies(vararg movie: Movie): List<Long>
     fun loadAllMovies(): Observable<List<Movie>>
-    fun updateMovieRating(movieName: String, rating: Int): Int
+    fun updateMovieRating(movieName: String, rating: Float): Int
 }
